Currently we detect the specific case that an event has no associated publisher metadata, and in that case return the event un-augmented. However in other cases such as OpenPublisherMetadata we error out of the whole routine.
It turns out that other buggy applications are easily able to poison the Event
Log in ways that cause OpenPublisherMetadata to fail. We've also seen other
failures that aren't obviously explained by this feature of the Event Log.

Either way, let's just return the data we _do_ have, so that one poisoned event
can't prevent us from reading subsequent events that show up in the
subscription.

PiperOrigin-RevId: 334855976