Don't bail out of GetRenderedEvents even if rendering fails entirely
In commit 4a3524d1a285 I changed GetRenderedEvents to log and soldier on when
it gets an error from RenderFormattedMessageXML. I was forgetting that we
_also_ experience errors from RenderFragment. We should treat them the same to
unblock these hosts that appear to have poisoned logs.

NB for RenderFormattedMessageXML failures, we print the fragment XML -
hopefully this will let us track down the source of the poisoning. We don't
have the same info for RenderFragment failures. I propose we go ahead anyway,
and fix the RenderFormattedMessageXML failures; we can then further investigate
the RenderFragment failures if they don't go away.

PiperOrigin-RevId: 335444819